refactoring: Remove domain logic
